Moved ui files into a separate location
