+++ ui_tests/collapsed_lines.txt
