+++ ui_tests/cube_colored.off
